restructured files for init values
