--- WebApp.Logger.Test/Extensions/ObjectInitExtension.cs
@@ -1,37 +1,0 @@
-﻿using MongoDB.Bson;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using WebApp.Common.Serialize;
-
-namespace WebApp.Logger.Test.Extensions
-{
-    public static class ObjectInitExtension
-    {
-        public static object GetErrorObject()
-        {
-            string json = "{\r\n  \"userId\": null,\r\n  \"application\": \"WebApp.Service\",\r\n  \"ipAddress\": \"::1\",\r\n  \"version\": \"https\",\r\n  \"host\": \"localhost:44373\",\r\n  \"url\": \"POST https://localhost:44373/api/Account/user/add\",\r\n  \"source\": \"WebApp.Service\",\r\n  \"form\": \"\",\r\n  \"body\": {\r\n    \"userName\": \"string\",\r\n    \"email\": \"string\",\r\n    \"firstname\": \"string\",\r\n    \"lastname\": \"string\",\r\n    \"password\": \"string\"\r\n  },\r\n  \"response\": null,\r\n  \"requestHeaders\": {\r\n    \"Accept\": [\r\n      \"*/*\"\r\n    ],\r\n    \"Accept-Encoding\": [\r\n      \"gzip, deflate, br\"\r\n    ],\r\n    \"Accept-Language\": [\r\n      \"en-US,en;q=0.9\"\r\n    ],\r\n    \"Connection\": [\r\n      \"close\"\r\n    ],\r\n    \"Content-Length\": [\r\n      \"99\"\r\n    ],\r\n    \"Content-Type\": [\r\n      \"application/json-patch+json\"\r\n    ],\r\n    \"Host\": [\r\n      \"localhost:44373\"\r\n    ],\r\n    \"Referer\": [\r\n      \"https://localhost:44373/swagger/index.html\"\r\n    ],\r\n    \"User-Agent\": [\r\n      \"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36\"\r\n    ],\r\n    \"sec-ch-ua\": [\r\n      \"\\\"Not?A_Brand\\\";v=\\\"8\\\", \\\"Chromium\\\";v=\\\"108\\\", \\\"Google Chrome\\\";v=\\\"108\\\"\"\r\n    ],\r\n    \"sec-ch-ua-mobile\": [\r\n      \"?0\"\r\n    ],\r\n    \"sec-ch-ua-platform\": [\r\n      \"\\\"Windows\\\"\"\r\n    ],\r\n    \"origin\": [\r\n      \"https://localhost:44373\"\r\n    ],\r\n    \"sec-fetch-site\": [\r\n      \"same-origin\"\r\n    ],\r\n    \"sec-fetch-mode\": [\r\n      \"cors\"\r\n    ],\r\n    \"sec-fetch-dest\": [\r\n      \"empty\"\r\n    ]\r\n  },\r\n  \"responseHeaders\": {\r\n    \"Accept\": [\r\n      \"*/*\"\r\n    ],\r\n    \"Accept-Encoding\": [\r\n      \"gzip, deflate, br\"\r\n    ],\r\n    \"Accept-Language\": [\r\n      \"en-US,en;q=0.9\"\r\n    ],\r\n    \"Connection\": [\r\n      \"close\"\r\n    ],\r\n    \"Content-Length\": [\r\n      \"99\"\r\n    ],\r\n    \"Content-Type\": [\r\n      \"application/json-patch+json\"\r\n    ],\r\n    \"Host\": [\r\n      \"localhost:44373\"\r\n    ],\r\n    \"Referer\": [\r\n      \"https://localhost:44373/swagger/index.html\"\r\n    ],\r\n    \"User-Agent\": [\r\n      \"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36\"\r\n    ],\r\n    \"sec-ch-ua\": [\r\n      \"\\\"Not?A_Brand\\\";v=\\\"8\\\", \\\"Chromium\\\";v=\\\"108\\\", \\\"Google Chrome\\\";v=\\\"108\\\"\"\r\n    ],\r\n    \"sec-ch-ua-mobile\": [\r\n      \"?0\"\r\n    ],\r\n    \"sec-ch-ua-platform\": [\r\n      \"\\\"Windows\\\"\"\r\n    ],\r\n    \"origin\": [\r\n      \"https://localhost:44373\"\r\n    ],\r\n    \"sec-fetch-site\": [\r\n      \"same-origin\"\r\n    ],\r\n    \"sec-fetch-mode\": [\r\n      \"cors\"\r\n    ],\r\n    \"sec-fetch-dest\": [\r\n      \"empty\"\r\n    ]\r\n  },\r\n  \"errorCode\": \"AE\",\r\n  \"scheme\": \"https\",\r\n  \"traceId\": \"400000d7-0002-f900-b63f-84710c7967bb\",\r\n  \"proctocol\": \"HTTP/2\",\r\n  \"errors\": [\r\n    {\r\n      \"code\": \"PasswordRequiresDigit\",\r\n      \"description\": \"Passwords must have at least one digit ('0'-'9').\"\r\n    }\r\n  ],\r\n  \"statusCode\": 200,\r\n  \"appStatusCode\": \"AP1300E\",\r\n  \"message\": \"User already exists with name string\",\r\n  \"messageDetails\": \"User already exists with name string\",\r\n  \"stackTrace\": \"   at WebApp.Service.Services.Accounts.UserService.AddUserAsync(User user, String password) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp.Service\\\\Services\\\\Accounts\\\\UserService.cs:line 92\\r\\n   at WebApp.Controllers.Auths.AccountController.AddUser(UserVm model) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp\\\\Controllers\\\\Auths\\\\AccountController.cs:line 52\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ActionMethodExecutor.TaskOfIActionResultExecutor.Execute(IActionResultTypeMapper mapper, ObjectMethodExecutor executor, Object controller, Object[] arguments)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.<InvokeActionMethodAsync>g__Awaited|12_0(ControllerActionInvoker invoker, ValueTask`1 actionResultValueTask)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.<InvokeNextActionFilterAsync>g__Awaited|10_0(ControllerActionInvoker invoker, Task lastTask, State next, Scope scope, Object state, Boolean isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.Rethrow(ActionExecutedContextSealed context)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.Next(State& next, Scope& scope, Object& state, Boolean& isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.<InvokeInnerFilterAsync>g__Awaited|13_0(ControllerActionInvoker invoker, Task lastTask, State next, Scope scope, Object state, Boolean isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ResourceInvoker.<InvokeFilterPipelineAsync>g__Awaited|19_0(ResourceInvoker invoker, Task lastTask, State next, Scope scope, Object state, Boolean isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ResourceInvoker.<InvokeAsync>g__Logged|17_1(ResourceInvoker invoker)\\r\\n   at Microsoft.AspNetCore.Routing.EndpointMiddleware.<Invoke>g__AwaitRequestTask|6_0(Endpoint endpoint, Task requestTask, ILogger logger)\\r\\n   at Microsoft.AspNetCore.Session.SessionMiddleware.Invoke(HttpContext context)\\r\\n   at Microsoft.AspNetCore.Session.SessionMiddleware.Invoke(HttpContext context)\\r\\n   at WebApp.Logger.Middlewares.HttpRequestMiddleware.InvokeAsync(HttpContext context, IServiceProvider _serviceProvider, IRouteLogRepository RouteLogRepository) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp.Logger\\\\Middlewares\\\\HttpRequestMiddleware.cs:line 56\\r\\n   at WebApp.Logger.Middlewares.ExceptionMiddleware.InvokeAsync(HttpContext context, IExceptionLogRepository exceptionLogRepository) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp.Logger\\\\Middlewares\\\\ExceptionMiddleware.cs:line 53\",\r\n  \"createdDateUtc\": null\r\n}";
-            return json.ToModel<object>();
-        }
-       
-        public static object GetAuditObject()
-        {
-            string json = "{\r\n  \"userId\": 0,\r\n  \"traceId\": \"400001b9-0000-fc00-b63f-84710c7967bb\",\r\n  \"type\": \"Update\",\r\n  \"schemaName\": \"dbo\",\r\n  \"tableName\": \"BlogEntity\",\r\n  \"dateTime\": \"2023-01-09T19:39:49.5268842+06:00\",\r\n  \"primaryKey\": {\r\n    \"id\": 307\r\n  },\r\n  \"oldValues\": {\r\n    \"createdDateUtc\": \"2023-01-09T19:39:09.6025626+06:00\",\r\n    \"description\": \"My blog description\",\r\n    \"motto\": \"Blog Motto\",\r\n    \"name\": \"My Blog\",\r\n    \"updatedBy\": 0,\r\n    \"updatedDateUtc\": null\r\n  },\r\n  \"newValues\": {\r\n    \"createdDateUtc\": \"0001-01-01T06:00:00+06:00\",\r\n    \"description\": \"Updated My blog description\",\r\n    \"motto\": \"Updated Blog Motto\",\r\n    \"name\": \"Updated My Blog\",\r\n    \"updatedBy\": 0,\r\n    \"updatedDateUtc\": \"2023-01-09T19:39:49.4912424+06:00\"\r\n  },\r\n  \"affectedColumns\": [\r\n    \"CreatedDateUtc\",\r\n    \"Description\",\r\n    \"Motto\",\r\n    \"Name\",\r\n    \"UpdatedDateUtc\"\r\n  ],\r\n  \"createdBy\": 0,\r\n  \"createdDateUtc\": \"2023-01-09T13:39:49.5268857Z\",\r\n  \"updatedBy\": 0,\r\n  \"updatedDateUtc\": null\r\n}";
-            return json.ToModel<object>();
-        }
-
-        public static object GetSqlObject()
-        {
-            string json = "{\r\n  \"userId\": null,\r\n  \"applicationName\": \"\",\r\n  \"ipAddress\": \"::1\",\r\n  \"version\": \"\",\r\n  \"host\": \"localhost:44373\",\r\n  \"url\": \"https://localhost:44373/api/Blog/edit/307\",\r\n  \"source\": \"Query\",\r\n  \"scheme\": \"https\",\r\n  \"traceId\": \"400001b9-0000-fc00-b63f-84710c7967bb\",\r\n  \"protocol\": \"HTTP/2\",\r\n  \"urlReferrer\": \"\",\r\n  \"area\": \"\",\r\n  \"controllerName\": \"Blog\",\r\n  \"actionName\": \"Edit\",\r\n  \"className\": \"\",\r\n  \"methodName\": \"\",\r\n  \"connection\": {\r\n    \"database\": \"DotnetLoggerWrapper\",\r\n    \"dataSource\": \"localhost\",\r\n    \"serverVersion\": \"15.00.2000\",\r\n    \"connectionId\": \"e4e77e9d-67a3-4ffd-9f6d-4e30e197ac3c\",\r\n    \"connectionTimeout\": 15\r\n  },\r\n  \"command\": {\r\n    \"commandTimeout\": 30,\r\n    \"commandType\": \"Text\"\r\n  },\r\n  \"event\": {\r\n    \"name\": \"Microsoft.EntityFrameworkCore.Database.Command.CommandExecuted\"\r\n  },\r\n  \"queryType\": \"Text\",\r\n  \"query\": \"SET NOCOUNT ON;\\r\\nUPDATE [Blogs] SET [CreatedBy] = @p0, [CreatedDateUtc] = @p1, [Description] = @p2, [Motto] = @p3, [Name] = @p4, [UpdatedBy] = @p5, [UpdatedDateUtc] = @p6\\r\\nWHERE [Id] = @p7;\\r\\nSELECT @@ROWCOUNT;\\r\\n\\r\\n\",\r\n  \"response\": null,\r\n  \"duration\": 14.6849,\r\n  \"message\": null,\r\n  \"createdDateUtc\": \"2023-01-09T13:39:49.5655732Z\"\r\n}";
-            return json.ToModel<object>();
-        }
-
-        public static object GetRequestObject()
-        {
-            string json = "{\r\n  \"userId\": null,\r\n  \"application\": null,\r\n  \"ipAddress\": \"127.0.0.1\",\r\n  \"version\": \"https\",\r\n  \"host\": \"localhost:44373\",\r\n  \"url\": \"POST https://localhost:44373/api/Blog/add\",\r\n  \"source\": null,\r\n  \"form\": \"\",\r\n  \"body\": {\r\n           \"name\":\"My Blog\",\r\n           \"description\":\"My blog description\",\r\n           \"motto\":\"Blog Motto\",\r\n           \"posts\":null,\r\n           \"id\":0},\r\n  \"response\": {\r\n                \"statusCode\":200,\r\n                \"appStatusCode\":\"AP1300E\",\r\n                \"isSuccess\":true,\r\n                \"message\":null,\r\n                \"data\":{\"name\":\"My Blog\",\"description\":\"My blog description\",\"motto\":\"Blog Motto\",\"posts\":[],\"id\":310},\"errors\":null},\r\n  \"requestHeaders\": {\r\n                \"Accept\":[\"*/*\"],\r\n                \"Accept-Encoding\":[\"gzip, deflate, br\"],\r\n                \"Accept-Language\":[\"en-US,en;q=0.9\"],\r\n                \"Connection\":[\"close\"],\r\n                \"Content-Length\":[\"95\"],\r\n                \"Content-Type\":[\"application/json-patch\\\\u002Bjson\"],\r\n                \"Host\":[\"localhost:44373\"],\r\n                \"Referer\":[\"https://localhost:44373/swagger/index.html\"],\r\n                \"User-Agent\":[\"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36\"],\r\n                \"sec-ch-ua\":[\"\\\\u0022Not?A_Brand\\\\u0022;v=\\\\u00228\\\\u0022, \\\\u0022Chromium\\\\u0022;v=\\\\u0022108\\\\u0022, \\\\u0022Google Chrome\\\\u0022;v=\\\\u0022108\\\\u0022\"],\r\n                \"sec-ch-ua-mobile\":[\"?0\"],\r\n                \"sec-ch-ua-platform\":[\"\\\\u0022Windows\\\\u0022\"],\r\n                \"origin\":[\"https://localhost:44373\"],\r\n                \"sec-fetch-site\":[\"same-origin\"],\r\n                \"sec-fetch-mode\":[\"cors\"],\r\n                \"sec-fetch-dest\":[\"empty\"]},\r\n  \"responseHeaders\": {\r\n                \"Accept\":[\"*/*\"],\r\n                \"Accept-Encoding\":[\"gzip, deflate, br\"],\r\n                \"Accept-Language\":[\"en-US,en;q=0.9\"],\r\n                \"Connection\":[\"close\"],\r\n                \"Content-Length\":[\"95\"],\r\n                \"Content-Type\":[\"application/json-patch\\\\u002Bjson\"],\r\n                \"Host\":[\"localhost:44373\"],\r\n                \"Referer\":[\"https://localhost:44373/swagger/index.html\"],\r\n                \"User-Agent\":[\"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36\"],\r\n                \"sec-ch-ua\":[\"\\\\u0022Not?A_Brand\\\\u0022;v=\\\\u00228\\\\u0022, \\\\u0022Chromium\\\\u0022;v=\\\\u0022108\\\\u0022, \\\\u0022Google Chrome\\\\u0022;v=\\\\u0022108\\\\u0022\"],\r\n                \"sec-ch-ua-mobile\":[\"?0\"],\"sec-ch-ua-platform\":[\"\\\\u0022Windows\\\\u0022\"],\r\n                \"origin\":[\"https://localhost:44373\"],\r\n                \"sec-fetch-site\":[\"same-origin\"],\r\n                \"sec-fetch-mode\":[\"cors\"],\r\n                \"sec-fetch-dest\":[\"empty\"]},\r\n  \"scheme\": \"https\",\r\n  \"traceId\": \"40000133-0000-f400-b63f-84710c7967bb\",\r\n  \"proctocol\": \"HTTP/2\",\r\n  \"area\": null,\r\n  \"controllerName\": null,\r\n  \"actionName\": null,\r\n  \"executionDuration\": 0.0,\r\n  \"roleId\": null,\r\n  \"languageId\": null,\r\n  \"isFirstLogin\": null,\r\n  \"loggedInDateTimeUtc\": null,\r\n  \"loggedOutDateTimeUtc\": null,\r\n  \"loginStatus\": null,\r\n  \"pageAccessed\": null,\r\n  \"sessionId\": null,\r\n  \"urlReferrer\": null,\r\n  \"statusCode\": 200,\r\n  \"appStatusCode\": \"*****\",\r\n  \"createdDateUtc\": null\r\n}";
-            return json.ToModel<object>();
-        }
-    }
-}
