+++ WebApp.Logger.Test/InitValues/ErrorObjectInit.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using WebApp.Logger.Extensions;
+
+namespace WebApp.Logger.Test.InitValues
+{
+    public static class ErrorObjectInit
+    {
+        public static object GetErrorObject()
+        {
+            string json = "{\r\n  \"userId\": null,\r\n  \"application\": \"WebApp.Service\",\r\n  \"ipAddress\": \"::1\",\r\n  \"version\": \"https\",\r\n  \"host\": \"localhost:44373\",\r\n  \"url\": \"POST https://localhost:44373/api/Account/user/add\",\r\n  \"source\": \"WebApp.Service\",\r\n  \"form\": \"\",\r\n  \"body\": {\r\n    \"userName\": \"string\",\r\n    \"email\": \"string\",\r\n    \"firstname\": \"string\",\r\n    \"lastname\": \"string\",\r\n    \"password\": \"string\"\r\n  },\r\n  \"response\": null,\r\n  \"requestHeaders\": {\r\n    \"Accept\": [\r\n      \"*/*\"\r\n    ],\r\n    \"Accept-Encoding\": [\r\n      \"gzip, deflate, br\"\r\n    ],\r\n    \"Accept-Language\": [\r\n      \"en-US,en;q=0.9\"\r\n    ],\r\n    \"Connection\": [\r\n      \"close\"\r\n    ],\r\n    \"Content-Length\": [\r\n      \"99\"\r\n    ],\r\n    \"Content-Type\": [\r\n      \"application/json-patch+json\"\r\n    ],\r\n    \"Host\": [\r\n      \"localhost:44373\"\r\n    ],\r\n    \"Referer\": [\r\n      \"https://localhost:44373/swagger/index.html\"\r\n    ],\r\n    \"User-Agent\": [\r\n      \"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36\"\r\n    ],\r\n    \"sec-ch-ua\": [\r\n      \"\\\"Not?A_Brand\\\";v=\\\"8\\\", \\\"Chromium\\\";v=\\\"108\\\", \\\"Google Chrome\\\";v=\\\"108\\\"\"\r\n    ],\r\n    \"sec-ch-ua-mobile\": [\r\n      \"?0\"\r\n    ],\r\n    \"sec-ch-ua-platform\": [\r\n      \"\\\"Windows\\\"\"\r\n    ],\r\n    \"origin\": [\r\n      \"https://localhost:44373\"\r\n    ],\r\n    \"sec-fetch-site\": [\r\n      \"same-origin\"\r\n    ],\r\n    \"sec-fetch-mode\": [\r\n      \"cors\"\r\n    ],\r\n    \"sec-fetch-dest\": [\r\n      \"empty\"\r\n    ]\r\n  },\r\n  \"responseHeaders\": {\r\n    \"Accept\": [\r\n      \"*/*\"\r\n    ],\r\n    \"Accept-Encoding\": [\r\n      \"gzip, deflate, br\"\r\n    ],\r\n    \"Accept-Language\": [\r\n      \"en-US,en;q=0.9\"\r\n    ],\r\n    \"Connection\": [\r\n      \"close\"\r\n    ],\r\n    \"Content-Length\": [\r\n      \"99\"\r\n    ],\r\n    \"Content-Type\": [\r\n      \"application/json-patch+json\"\r\n    ],\r\n    \"Host\": [\r\n      \"localhost:44373\"\r\n    ],\r\n    \"Referer\": [\r\n      \"https://localhost:44373/swagger/index.html\"\r\n    ],\r\n    \"User-Agent\": [\r\n      \"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36\"\r\n    ],\r\n    \"sec-ch-ua\": [\r\n      \"\\\"Not?A_Brand\\\";v=\\\"8\\\", \\\"Chromium\\\";v=\\\"108\\\", \\\"Google Chrome\\\";v=\\\"108\\\"\"\r\n    ],\r\n    \"sec-ch-ua-mobile\": [\r\n      \"?0\"\r\n    ],\r\n    \"sec-ch-ua-platform\": [\r\n      \"\\\"Windows\\\"\"\r\n    ],\r\n    \"origin\": [\r\n      \"https://localhost:44373\"\r\n    ],\r\n    \"sec-fetch-site\": [\r\n      \"same-origin\"\r\n    ],\r\n    \"sec-fetch-mode\": [\r\n      \"cors\"\r\n    ],\r\n    \"sec-fetch-dest\": [\r\n      \"empty\"\r\n    ]\r\n  },\r\n  \"errorCode\": \"AE\",\r\n  \"scheme\": \"https\",\r\n  \"traceId\": \"400000d7-0002-f900-b63f-84710c7967bb\",\r\n  \"proctocol\": \"HTTP/2\",\r\n  \"errors\": [\r\n    {\r\n      \"code\": \"PasswordRequiresDigit\",\r\n      \"description\": \"Passwords must have at least one digit ('0'-'9').\"\r\n    }\r\n  ],\r\n  \"statusCode\": 200,\r\n  \"appStatusCode\": \"AP1300E\",\r\n  \"message\": \"User already exists with name string\",\r\n  \"messageDetails\": \"User already exists with name string\",\r\n  \"stackTrace\": \"   at WebApp.Service.Services.Accounts.UserService.AddUserAsync(User user, String password) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp.Service\\\\Services\\\\Accounts\\\\UserService.cs:line 92\\r\\n   at WebApp.Controllers.Auths.AccountController.AddUser(UserVm model) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp\\\\Controllers\\\\Auths\\\\AccountController.cs:line 52\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ActionMethodExecutor.TaskOfIActionResultExecutor.Execute(IActionResultTypeMapper mapper, ObjectMethodExecutor executor, Object controller, Object[] arguments)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.<InvokeActionMethodAsync>g__Awaited|12_0(ControllerActionInvoker invoker, ValueTask`1 actionResultValueTask)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.<InvokeNextActionFilterAsync>g__Awaited|10_0(ControllerActionInvoker invoker, Task lastTask, State next, Scope scope, Object state, Boolean isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.Rethrow(ActionExecutedContextSealed context)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.Next(State& next, Scope& scope, Object& state, Boolean& isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker.<InvokeInnerFilterAsync>g__Awaited|13_0(ControllerActionInvoker invoker, Task lastTask, State next, Scope scope, Object state, Boolean isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ResourceInvoker.<InvokeFilterPipelineAsync>g__Awaited|19_0(ResourceInvoker invoker, Task lastTask, State next, Scope scope, Object state, Boolean isCompleted)\\r\\n   at Microsoft.AspNetCore.Mvc.Infrastructure.ResourceInvoker.<InvokeAsync>g__Logged|17_1(ResourceInvoker invoker)\\r\\n   at Microsoft.AspNetCore.Routing.EndpointMiddleware.<Invoke>g__AwaitRequestTask|6_0(Endpoint endpoint, Task requestTask, ILogger logger)\\r\\n   at Microsoft.AspNetCore.Session.SessionMiddleware.Invoke(HttpContext context)\\r\\n   at Microsoft.AspNetCore.Session.SessionMiddleware.Invoke(HttpContext context)\\r\\n   at WebApp.Logger.Middlewares.HttpRequestMiddleware.InvokeAsync(HttpContext context, IServiceProvider _serviceProvider, IRouteLogRepository RouteLogRepository) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp.Logger\\\\Middlewares\\\\HttpRequestMiddleware.cs:line 56\\r\\n   at WebApp.Logger.Middlewares.ExceptionMiddleware.InvokeAsync(HttpContext context, IExceptionLogRepository exceptionLogRepository) in C:\\\\BS-CS-Basic-Team\\\\AspNetCoreLogger\\\\WebApp.Logger\\\\Middlewares\\\\ExceptionMiddleware.cs:line 53\",\r\n  \"createdDateUtc\": null\r\n}";
+            return json.ToModel<object>();
+        }
+    }
+}
