+++ WebApp.Logger/Endpoints/LoggerWrapperController.cs
@@ -20,13 +20,11 @@ namespace WebApp.Logger
         private readonly IRouteLogRepository _routeLogRepository;
         private readonly IExceptionLogRepository _exceptionLogRepository;
         private readonly IAuditLogRepository _auditLogRepository;
-        //private readonly IServiceProvider _serviceProvider;
         private readonly LogOption _logOption;
 
         public LoggerWrapperController(IRouteLogRepository routeLogRepository,
             IExceptionLogRepository exceptionLogRepository,
             IAuditLogRepository auditLogRepository,
-            //IServiceProvider serviceProvider,
             ISqlLogRepository sqlLogRepository,
             IOptions<LogOption> logOption)
         {
@@ -35,18 +33,11 @@ namespace WebApp.Logger
             _auditLogRepository = auditLogRepository;
             _sqlLogRepository = sqlLogRepository;
             _logOption = logOption.Value;
-
-            //var factory = new ProviderFactory(_serviceProvider);
-
-            //var providerType = _logOption.ProviderType;
-            //_loggerWrapper = factory.Build(providerType);
         }
 
         [HttpGet("routes")]
         public async Task<IActionResult> GetRouteLogsAsync(int pageIndex = CommonVariables.pageIndex, int pageSize = CommonVariables.pageSize, string searchText = null)
         {
-            //var res = await _loggerWrapper.Request.GetPageAsync(new DapperPager(pageIndex, pageSize));
-            
             var res = await _routeLogRepository.GetPageAsync(new DapperPager(pageIndex, pageSize));
 
             return new OkResponse(res);
@@ -58,8 +49,6 @@ namespace WebApp.Logger
             string continuationToken = null,
             string searchText = null)
         {
-            //var res = await _loggerWrapper.Audit.GetPageAsync(new DapperPager(pageIndex, continuationToken, pageSize));
-            
             var res = await _auditLogRepository.GetPageAsync(new DapperPager(pageIndex, continuationToken, pageSize));
 
             return new OkResponse(res);
@@ -68,8 +57,6 @@ namespace WebApp.Logger
         [HttpGet("exceptions")]
         public async Task<IActionResult> GetExceptionLogssAsync(int pageIndex = CommonVariables.pageIndex, int pageSize = CommonVariables.pageSize, string searchText = null)
         {
-            //var res = await _loggerWrapper.Error.GetPageAsync(new DapperPager(pageIndex, pageSize));
-
             var res = await _exceptionLogRepository.GetPageAsync(new DapperPager(pageIndex, pageSize));
 
             return new OkResponse(res);
@@ -78,8 +65,6 @@ namespace WebApp.Logger
         [HttpGet("sqls")]
         public async Task<IActionResult> GetSqlLogssAsync(int pageIndex = CommonVariables.pageIndex, int pageSize = CommonVariables.pageSize, string searchText = null)
         {
-            //var res = await _loggerWrapper.Sql.GetPageAsync(new DapperPager(pageIndex, pageSize));
-
             var res = await _sqlLogRepository.GetPageAsync(new DapperPager(pageIndex, pageSize));
 
             return new OkResponse(res);
