+++ WebApp/Controllers/Logs/LogController.cs
@@ -31,18 +31,11 @@ namespace WebApp.Controllers.Logs
             _exceptionLogRepository = exceptionLogRepository;
             _auditLogRepository = auditLogRepository;
             _sqlLogRepository = sqlLogRepository;
-
-            //var factory = new ProviderFactory(_serviceProvider);
-
-            //var providerType = _logOption.ProviderType;
-            //_loggerWrapper = factory.Build(providerType);
         }
 
         [HttpGet("routes")]
         public async Task<IActionResult> GetRouteLogsAsync(int pageIndex = CommonVariables.pageIndex, int pageSize = CommonVariables.pageSize, string searchText = null)
         {
-            //var res = await _loggerWrapper.Request.GetPageAsync(new DapperPager(pageIndex, pageSize));
-
             var res = await _routeLogRepository.GetPageAsync(new DapperPager(pageIndex, pageSize));
 
             return new OkResponse(res);
@@ -54,8 +47,6 @@ namespace WebApp.Controllers.Logs
             string continuationToken = null,
             string searchText = null)
         {
-            //var res = await _loggerWrapper.Audit.GetPageAsync(new DapperPager(pageIndex, continuationToken, pageSize));
-
             var res = await _auditLogRepository.GetPageAsync(new DapperPager(pageIndex, continuationToken, pageSize));
 
             return new OkResponse(res);
@@ -64,8 +55,6 @@ namespace WebApp.Controllers.Logs
         [HttpGet("exceptions")]
         public async Task<IActionResult> GetExceptionLogssAsync(int pageIndex = CommonVariables.pageIndex, int pageSize = CommonVariables.pageSize, string searchText = null)
         {
-            //var res = await _loggerWrapper.Error.GetPageAsync(new DapperPager(pageIndex, pageSize));
-
             var res = await _exceptionLogRepository.GetPageAsync(new DapperPager(pageIndex, pageSize));
 
             return new OkResponse(res);
@@ -74,8 +63,6 @@ namespace WebApp.Controllers.Logs
         [HttpGet("sqls")]
         public async Task<IActionResult> GetSqlLogssAsync(int pageIndex = CommonVariables.pageIndex, int pageSize = CommonVariables.pageSize, string searchText = null)
         {
-            //var res = await _loggerWrapper.Sql.GetPageAsync(new DapperPager(pageIndex, pageSize));
-
             var res = await _sqlLogRepository.GetPageAsync(new DapperPager(pageIndex, pageSize));
 
             return new OkResponse(res);
